+++ Sources/Sentry/SentryMechanism.m
@@ -1,4 +1,5 @@
 #import "SentryMechanism.h"
+#import "NSDictionary+SentrySanitize.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -20,7 +21,7 @@ NS_ASSUME_NONNULL_BEGIN
     [serializedData setValue:self.handled forKey:@"handled"];
     [serializedData setValue:self.desc forKey:@"description"];
     [serializedData setValue:self.meta forKey:@"meta"];
-    [serializedData setValue:self.data forKey:@"data"];
+    [serializedData setValue:[self.data sentry_sanitize] forKey:@"data"];
     [serializedData setValue:self.helpLink forKey:@"help_link"];
 
     return serializedData;